requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
langgraph
requests
yt-dlp
opencv-python
numpy
